+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 ## [Unreleased]
 
+## [0.8.0]
+
 ### Changed
 
 - **Internal**: Consolidated Component and Asset registries into a single `ComponentRegistry`.
@@ -154,7 +156,7 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 
 - Josh Thomas <josh@joshthomas.dev> (maintainer)
 
-[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.7.2...HEAD
+[unreleased]: https://github.com/joshuadavidthomas/django-bird/compare/v0.8.0...HEAD
 [0.1.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.0
 [0.1.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.1.1
 [0.2.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.2.0
@@ -167,3 +169,4 @@ and this project attempts to adhere to [Semantic Versioning](https://semver.org/
 [0.7.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.7.0
 [0.7.1]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.7.1
 [0.7.2]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.7.2
+[0.8.0]: https://github.com/joshuadavidthomas/django-bird/releases/tag/v0.8.0